[REFACTOR] bookmark 라우터 분리 #74
